Update bib_sources.csv
Bringing bib_sources list in line with config.bib_sources in the database. I made assumptions as to whether the De Gruyter trial was a subscription, and whether the Course materials module was local/purchased. I checked with Melissa regarding both of them and will update again if I am incorrect in my info.